New Classes:
StockDay: Acts as the template for holding information from one day of a
stock.
	Most functions are assessors and mutators but one 	noteworthy function
	is getDateNumber() since returns the date as a 	number (which I used
	for date comparisons)

StockList: acts as the template for a stock over series of dates. The
constructor with a String parameter uses the parameter to load in
information, but only in the location and form that the original
IDownloadData provides.
The findEngulfPattern and sortDateHighToLow functions should be
self-explanatory  .